pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cargo-compliance-env"
version = "0.1.0"
description = "Bilateral Cargo Compliance & Regulatory RL Environment"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Joel Jose" }]
dependencies = [
    "openenv-core>=0.2.0",
    "fastapi",
    "uvicorn",
    "openai",
    "httpx",
    "requests",
    "groq",
    "python-dotenv",
    "tenacity",
    "pydantic"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Validator-friendly multi-mode entry points
server = "server.main:main"
agent = "inference:run"
inference = "inference:run"
# Backward-compatible aliases
cargo-env = "server.main:main"
cargo-agent = "inference:run"
i-agent = "i:run"

[tool.setuptools]
# Ensures both the server logic and your agent script are discoverable
packages = ["server"]
py-modules = ["inference", "i"]

[tool.setuptools.package-data]
# Includes the final_dataset.json and openenv.yaml in the distribution
"*" = ["*.json", "*.yaml"]
